reorg types & ensure id exists
