+++ src/contexts/Connect/index.tsx
@@ -6,12 +6,15 @@ import { DappName } from 'consts';
 import { useApi } from 'contexts/Api';
 import {
   ConnectContextInterface,
-  ExtensionAccount,
   ExternalAccount,
   ImportedAccount,
 } from 'contexts/Connect/types';
 import { useExtensions } from 'contexts/Extensions';
-import { Extension, ExtensionInteface } from 'contexts/Extensions/types';
+import {
+  Extension,
+  ExtensionAccount,
+  ExtensionInteface,
+} from 'contexts/Extensions/types';
 import React, { useEffect, useRef, useState } from 'react';
 import { AnyApi, MaybeAccount } from 'types';
 import { clipAddress, localStorageOrDefault, setStateWithRef } from 'Utils';
@@ -220,60 +223,67 @@ export const ConnectProvider = ({
     let i = 0;
     extensions.forEach(async (e: Extension) => {
       i++;
-      const { id, enable } = e;
 
-      // if extension is found locally, subscribe to accounts
-      if (extensionIsLocal(id)) {
-        try {
-          // summons extension popup
-          const extension: ExtensionInteface = await enable(DappName);
+      // ensure the extension carries an `id` property
+      const id = e?.id ?? undefined;
 
-          if (extension !== undefined) {
-            const unsub = (await extension.accounts.subscribe(
-              (injected: ExtensionAccount[]) => {
-                if (injected) {
-                  injected = handleImportExtension(
-                    id,
-                    accountsRef.current,
-                    extension,
-                    injected,
-                    forgetAccounts
-                  );
-                  // store active wallet account if found in this extension
-                  if (!activeWalletAccount) {
-                    activeWalletAccount = getActiveExtensionAccount(injected);
-                  }
-                  // set active account for network on final extension
-                  if (i === total && activeAccountRef.current === null) {
-                    connectActiveExtensionAccount(
-                      activeWalletAccount,
-                      connectToAccount
+      if (id) {
+        // if extension is found locally, subscribe to accounts
+        if (extensionIsLocal(id)) {
+          try {
+            // attempt to get extension `enable` property
+            const { enable } = e;
+
+            // summons extension popup
+            const extension: ExtensionInteface = await enable(DappName);
+
+            if (extension !== undefined) {
+              const unsub = (await extension.accounts.subscribe(
+                (injected: ExtensionAccount[]) => {
+                  if (injected) {
+                    injected = handleImportExtension(
+                      id,
+                      accountsRef.current,
+                      extension,
+                      injected,
+                      forgetAccounts
                     );
-                  }
-                  // concat accounts and store
-                  if (injected.length) {
-                    setStateWithRef(
-                      [...accountsRef.current].concat(injected),
-                      setAccounts,
-                      accountsRef
-                    );
+                    // store active wallet account if found in this extension
+                    if (!activeWalletAccount) {
+                      activeWalletAccount = getActiveExtensionAccount(injected);
+                    }
+                    // set active account for network on final extension
+                    if (i === total && activeAccountRef.current === null) {
+                      connectActiveExtensionAccount(
+                        activeWalletAccount,
+                        connectToAccount
+                      );
+                    }
+                    // concat accounts and store
+                    if (injected.length) {
+                      setStateWithRef(
+                        [...accountsRef.current].concat(injected),
+                        setAccounts,
+                        accountsRef
+                      );
+                    }
                   }
                 }
-              }
-            )) as () => void;
+              )) as () => void;
 
-            // update context state
-            setStateWithRef(
-              [...unsubscribeRef.current].concat({
-                key: id,
-                unsub,
-              }),
-              setUnsubscribe,
-              unsubscribeRef
-            );
+              // update context state
+              setStateWithRef(
+                [...unsubscribeRef.current].concat({
+                  key: id,
+                  unsub,
+                }),
+                setUnsubscribe,
+                unsubscribeRef
+              );
+            }
+          } catch (err) {
+            handleExtensionError(id, String(err));
           }
-        } catch (err) {
-          handleExtensionError(id, String(err));
         }
       }
 
@@ -293,53 +303,60 @@ export const ConnectProvider = ({
   const connectExtensionAccounts = async (e: Extension) => {
     const keyring = new Keyring();
     keyring.setSS58Format(network.ss58);
-    const { id, enable } = e;
 
-    try {
-      // summons extension popup
-      const extension: ExtensionInteface = await enable(DappName);
+    // ensure the extension carries an `id` property
+    const id = e?.id ?? undefined;
 
-      if (extension !== undefined) {
-        // subscribe to accounts
-        const unsub = (await extension.accounts.subscribe(
-          (injected: ExtensionAccount[]) => {
-            if (injected) {
-              injected = handleImportExtension(
-                id,
-                accountsRef.current,
-                extension,
-                injected,
-                forgetAccounts
-              );
-              // set active account for network if not yet set
-              if (activeAccountRef.current === null) {
-                connectActiveExtensionAccount(
-                  getActiveExtensionAccount(injected),
-                  connectToAccount
+    if (id) {
+      try {
+        // attempt to get extension `enable` property
+        const { enable } = e;
+
+        // summons extension popup
+        const extension: ExtensionInteface = await enable(DappName);
+
+        if (extension !== undefined) {
+          // subscribe to accounts
+          const unsub = (await extension.accounts.subscribe(
+            (injected: ExtensionAccount[]) => {
+              if (injected) {
+                injected = handleImportExtension(
+                  id,
+                  accountsRef.current,
+                  extension,
+                  injected,
+                  forgetAccounts
+                );
+                // set active account for network if not yet set
+                if (activeAccountRef.current === null) {
+                  connectActiveExtensionAccount(
+                    getActiveExtensionAccount(injected),
+                    connectToAccount
+                  );
+                }
+                // concat accounts and store
+                setStateWithRef(
+                  [...accountsRef.current].concat(injected),
+                  setAccounts,
+                  accountsRef
                 );
               }
-              // concat accounts and store
-              setStateWithRef(
-                [...accountsRef.current].concat(injected),
-                setAccounts,
-                accountsRef
-              );
             }
-          }
-        )) as () => void;
+          )) as () => void;
 
-        // update context state
-        setStateWithRef(
-          [...unsubscribeRef.current].concat({
-            key: id,
-            unsub,
-          }),
-          setUnsubscribe,
-          unsubscribeRef
-        );
+          // update context state
+          setStateWithRef(
+            [...unsubscribeRef.current].concat({
+              key: id,
+              unsub,
+            }),
+            setUnsubscribe,
+            unsubscribeRef
+          );
+        }
+      } catch (err) {
+        handleExtensionError(id, String(err));
       }
-    } catch (err) {
-      handleExtensionError(id, String(err));
     }
   };
